requirements.txt
Django
Fabric
djangorestframework
ecdsa
ipython
paramiko
pycrypto
wsgiref
